# api/src/controllers/incidents.controller.ts
/**
 * Handlers for /v1/incidents. Zod request schemas are written in full; handler bodies are
 * stubbed (business logic lives in src/services/incidentService.ts, also stubbed).
 * Component: api/src/controllers
 */

import type { NextFunction, Request, Response } from "express";
import { z } from "zod";
import { INCIDENT_STATUSES } from "../types/incident.js";

export const listIncidentsQuerySchema = z.object({
  status: z.enum(INCIDENT_STATUSES).optional(),
  limit: z.coerce.number().int().positive().max(1000).optional(),
});

export const incidentIdParamsSchema = z.object({
  id: z.string().min(1),
});

export const updateIncidentStatusBodySchema = z.object({
  status: z.enum(INCIDENT_STATUSES),
});

/**
 * GET /v1/incidents
 * Input: query params validated by listIncidentsQuerySchema.
 * Output: 200 Incident[]
 * Error cases: 400 on invalid query params.
 */
export function listIncidents(req: Request, res: Response, next: NextFunction): void {
  // TODO(student): parse req.query, call incidentService.listIncidents(filter), respond 200.
  void req;
  void res;
  void next;
  throw new Error("incidents.controller.listIncidents not implemented");
}

/**
 * GET /v1/incidents/:id
 * Input: params validated by incidentIdParamsSchema.
 * Output: 200 Incident
 * Error cases: 404 if no incident with that id exists.
 */
export function getIncident(req: Request, res: Response, next: NextFunction): void {
  // TODO(student): parse req.params, call incidentService.getIncidentById(id), respond 200.
  void req;
  void res;
  void next;
  throw new Error("incidents.controller.getIncident not implemented");
}

/**
 * POST /v1/incidents/analyze
 * Input: none (operates on all currently-stored events).
 * Output: 201 Incident[] (newly created incidents from this run)
 * Error cases: 500 (ANALYZER_ERROR) if the analyzer process fails, times out, or returns
 * unparseable output.
 */
export function analyzeIncidents(req: Request, res: Response, next: NextFunction): void {
  // TODO(student): call incidentService.analyzeAndPersist(), respond 201 with the created
  // incidents.
  void req;
  void res;
  void next;
  throw new Error("incidents.controller.analyzeIncidents not implemented");
}

/**
 * PATCH /v1/incidents/:id/status
 * Input: params validated by incidentIdParamsSchema, body validated by
 * updateIncidentStatusBodySchema.
 * Output: 200 Incident (updated)
 * Error cases: 400 on invalid status value, 404 if no incident with that id exists.
 */
export function updateIncidentStatus(req: Request, res: Response, next: NextFunction): void {
  // TODO(student): parse req.params and req.body, call
  // incidentService.updateIncidentStatus(id, status), respond 200.
  void req;
  void res;
  void next;
  throw new Error("incidents.controller.updateIncidentStatus not implemented");
}

/**
 * DELETE /v1/incidents/:id
 * Input: params validated by incidentIdParamsSchema.
 * Output: 204 No Content
 * Error cases: 404 if no incident with that id exists.
 */
export function deleteIncident(req: Request, res: Response, next: NextFunction): void {
  // TODO(student): parse req.params, call incidentService.deleteIncident(id), respond 204.
  void req;
  void res;
  void next;
  throw new Error("incidents.controller.deleteIncident not implemented");
}